+++ package.json
@@ -33,7 +33,8 @@
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s --pkg package.json",
     "pre-publish": "node scripts/publish/publish.js && node scripts/publish/publish-scaffold.js",
     "publish:next": "bash scripts/publish/publish.sh -next",
-    "publish": "bash scripts/publish/publish.sh"
+    "publish": "bash scripts/publish/publish.sh",
+    "sync-scaffold-version": "node scripts/publish/sync-scaffold-version.js"
   },
   "dependencies": {
     "@angular/animations": "~7.2.0",
@@ -50,7 +51,7 @@
     "tslib": "^1.9.0",
     "zone.js": "~0.8.26",
     "@antv/data-set": "^0.10.1",
-    "@antv/g2": "^3.4.5",
+    "@antv/g2": "^3.4.8",
     "@antv/g2-plugin-slider": "^2.1.1",
     "@ngx-translate/core": "^11.0.1",
     "@stackblitz/sdk": "^1.3.0",
@@ -74,7 +75,7 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.0",
+    "@angular/cli": "~7.3.3",
     "@angular/compiler-cli": "~7.2.0",
     "@angular/language-service": "~7.2.0",
     "@types/node": "~8.9.4",
@@ -99,23 +100,23 @@
     "@types/extend": "^3.0.0",
     "@types/fs-extra": "^5.0.4",
     "@types/htmlparser2": "^3.7.31",
-    "@types/jszip": "^3.1.3",
-    "@types/mockjs": "^1.0.0",
+    "@types/jszip": "^3.1.5",
+    "@types/mockjs": "^1.0.2",
     "@types/parse5": "^5.0.0",
-    "karma-summary-reporter": "^1.5.2",
+    "karma-summary-reporter": "^1.6.0",
     "karma-parallel": "^0.3.1",
-    "tslint-config-prettier": "^1.17.0",
+    "tslint-config-prettier": "^1.18.0",
     "tslint-language-service": "^0.9.9",
-    "antd-theme-generator": "^1.1.4",
-    "chalk": "^2.4.1",
-    "codecov": "^3.1.0",
+    "antd-theme-generator": "^1.1.6",
+    "chalk": "^2.4.2",
+    "codecov": "^3.2.0",
     "conventional-changelog-cli": "^2.0.5",
     "domhandler": "^2.4.2",
     "editorconfig-tools": "^0.1.1",
     "fs-extra": "^7.0.0",
     "gh-pages": "^2.0.1",
     "htmlparser2": "^3.10.0",
-    "husky": "^1.1.0",
+    "husky": "^1.3.1",
     "jsonml.js": "^0.1.0",
     "klaw-sync": "^6.0.0",
     "less-plugin-clean-css": "^1.5.1",
@@ -123,7 +124,7 @@
     "mockjs": "^1.0.1-beta3",
     "mustache": "^3.0.0",
     "ng-packagr": "^4.7.1",
-    "tsickle": "^0.33.1",
+    "tsickle": "^0.34.3",
     "parse5": "^5.0.0",
     "prettier": "^1.16.1",
     "prettier-stylelint": "^0.4.2",
@@ -131,7 +132,7 @@
     "stream": "0.0.2",
     "stylelint": "^9.10.1",
     "stylelint-config-standard": "^18.2.0",
-    "stylelint-config-prettier": "^4.0.0",
+    "stylelint-config-prettier": "^5.0.0",
     "webpack-bundle-analyzer": "^3.0.3",
     "xlsx": "^0.14.0",
     "screenfull": "^4.0.0",
