+++ packages/abc/notice-icon/notice-icon.spec.ts
@@ -1,11 +1,11 @@
 import { Component, DebugElement, Injector, ViewChild } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { configureTestSuite, createTestContext } from '@delon/testing';
 import { en_US, zh_CN, DelonLocaleModule, DelonLocaleService } from '@delon/theme';
+import { NzIconTestModule } from 'ng-zorro-antd';
 
-import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { NoticeIconComponent } from './notice-icon.component';
 import { NoticeIconModule } from './notice-icon.module';
 import { NoticeItem } from './notice-icon.types';
@@ -18,7 +18,7 @@ describe('abc: notice-icon', () => {
 
   configureTestSuite(() => {
     injector = TestBed.configureTestingModule({
-      imports: [NoopAnimationsModule, NoticeIconModule, HttpClientTestingModule, DelonLocaleModule],
+      imports: [NoopAnimationsModule, NoticeIconModule, NzIconTestModule, DelonLocaleModule],
       declarations: [TestComponent],
     });
   });
