+++ scripts/publish/publish-scaffold.js
@@ -31,34 +31,10 @@ const log = {
 log.info('Starting publishing process...');
 
 fetchOlderVersions();
-fixVersion();
+execSync('npm run sync-scaffold-version', execSyncOptions);
 checkout();
 log.success(`Use commit message:`);
 log.success(`release(${nextVersion}): release ${nextVersion}`);
-
-function fixVersion() {
-  const packagePath = path.join(root, './package.json');
-  const json = fs.readJSONSync(packagePath);
-  json.version = nextVersion;
-  [
-    'abc',
-    'acl',
-    'auth',
-    'chart',
-    'cache',
-    'mock',
-    'form',
-    'theme',
-    'util',
-  ].forEach(v => {
-    json.dependencies[`@delon/${v}`] = `^${nextVersion}`;
-  });
-  json.devDependencies[`@delon/testing`] = `^${nextVersion}`;
-  json.devDependencies[`ng-alain`] = `^${nextVersion}`;
-  fs.writeJSONSync(packagePath, json, {
-    spaces: 2
-  });
-}
 
 function fetchOlderVersions() {
   log.info('Fetching older versions...');
